restructure components and rename `utils.py` to `common.py`  (#1119)
