+++ .readthedocs.yml
@@ -1,8 +1,10 @@
 version: 2
+sphinx:
+  configuration: doc/conf.py
 build:
   os: ubuntu-24.04
   tools:
-    python: "3.12"
+    python: "3.13"
   jobs:
     pre_system_dependencies:
     - git fetch --unshallow # Needed to get version tags
@@ -14,4 +16,3 @@ python:
     - docs
     - dev
     - cartopy
-    - gurobipy
